requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
defusedxml==0.5.0
Django==2.0.3
django-appconf==1.0.2
django-cleanup==2.1.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-unused-media==0.1.11
djangorestframework==3.7.7
djangorestframework-xml==1.3.0
factory-boy==2.10.0
Faker==0.8.11
idna==2.6
Jinja2==2.10
livereload==2.5.1
Markdown==2.6.11
MarkupSafe==1.0
mkdocs==0.17.3
Pillow==5.0.0
psycopg2==2.7.4
python-dateutil==2.7.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
tornado==4.5.3
urllib3==1.22
